Fix position_ids' dtype is int64 in CUDA graph runner.

+++ tensorrt_llm/_torch/models/modeling_qwen3.py
@@ -90,14 +90,12 @@ class Qwen3Attention(Attention):
         return q, k
 
     def apply_qk_norm_rope(self, qkv, position_ids):
-        #print("###########position_ids.shape", position_ids.shape)
         torch.ops.trtllm.fused_qk_norm_rope(
             qkv, self.num_heads, self.num_key_value_heads,
             self.num_key_value_heads, self.head_dim,
             self.q_norm.variance_epsilon, self.q_norm.weight,
             self.k_norm.weight, self.rotary_emb.rope_params.theta,
-            self.rotary_emb.is_neox,
-            position_ids.view(-1).int())
+            self.rotary_emb.is_neox, position_ids.view(-1))
         return qkv
